+++ src/Control/Eff/Concurrent/Api/Observer.hs
@@ -66,8 +66,8 @@ notifyObserver
   -> Server o
   -> Observation o
   -> Eff r ()
-notifyObserver px observer observed observation =
-  cast px observer (observationMessage observed observation)
+notifyObserver px observer observed observation = withFrozenCallStack
+  $ cast px observer (observationMessage observed observation)
 
 -- | Send the 'registerObserverMessage'
 registerObserver
@@ -76,8 +76,8 @@ registerObserver
   -> Server p
   -> Server o
   -> Eff r ()
-registerObserver px observer observed =
-  cast px observed (registerObserverMessage (SomeObserver observer))
+registerObserver px observer observed = withFrozenCallStack
+  $ cast px observed (registerObserverMessage (SomeObserver observer))
 
 -- | Send the 'forgetObserverMessage'
 forgetObserver
@@ -86,8 +86,8 @@ forgetObserver
   -> Server p
   -> Server o
   -> Eff r ()
-forgetObserver px observer observed =
-  cast px observed (forgetObserverMessage (SomeObserver observer))
+forgetObserver px observer observed = withFrozenCallStack
+  $ cast px observed (forgetObserverMessage (SomeObserver observer))
 
 -- | An existential wrapper around a 'Server' of an 'Observer'.
 -- Needed to support different types of observers to observe the
@@ -114,7 +114,7 @@ notifySomeObserver
   -> SomeObserver o
   -> Eff r ()
 notifySomeObserver px observed observation (SomeObserver observer) =
-  notifyObserver px observer observed observation
+  withFrozenCallStack $ notifyObserver px observer observed observation
 
 -- ** Manage 'Observers's
 
@@ -186,12 +186,13 @@ spawnCallbackObserver
      , Show (Observation o)
      , Observable o
      , Member (Logs LogMessage) q
+     , HasCallStack
      )
   => SchedulerProxy q
   -> (Server o -> Observation o -> Eff (Process q ': q) Bool)
   -> Eff r (Server (CallbackObserver o))
 spawnCallbackObserver px onObserve =
-  asServer @(CallbackObserver o) <$> spawn @r @q
+  withFrozenCallStack $ asServer @(CallbackObserver o) <$> spawn @r @q
     (do
       let loopUntil = serve
             px
@@ -217,9 +218,10 @@ spawnLoggingObserver
      , Show (Observation o)
      , Observable o
      , Member (Logs LogMessage) q
+     , HasCallStack
      )
   => SchedulerProxy q
   -> Eff r (Server (CallbackObserver o))
-spawnLoggingObserver px = spawnCallbackObserver
+spawnLoggingObserver px = withFrozenCallStack $ spawnCallbackObserver
   px
   (\s o -> logDebug (show s ++ " OBSERVED: " ++ show o) >> return True)