+++ src/Control/Eff/Concurrent/Process.hs
@@ -142,7 +142,7 @@ executeAndResume
    . (SetMember Process (Process q) r, HasCallStack)
   => Process q (ResumeProcess v)
   -> Eff r v
-executeAndResume processAction = do
+executeAndResume processAction = withFrozenCallStack $ do
   result <- send processAction
   case result of
     ResumeWith !value -> return value
@@ -157,7 +157,7 @@ executeAndCatch
   => SchedulerProxy q
   -> Eff r (ResumeProcess v)
   -> Eff r (Either String v)
-executeAndCatch px processAction = do
+executeAndCatch px processAction = withFrozenCallStack $ do
   result <- processAction
   case result of
     ResumeWith !value -> return (Right value)
@@ -186,7 +186,7 @@ yieldProcess
    . (HasCallStack, SetMember Process (Process q) r)
   => SchedulerProxy q
   -> Eff r ()
-yieldProcess _ = executeAndResume YieldProcess
+yieldProcess _ = withFrozenCallStack $ executeAndResume YieldProcess
 
 -- | Send a message to a process addressed by the 'ProcessId'.
 -- See 'SendMessage'.
@@ -197,7 +197,8 @@ sendMessage
   -> ProcessId
   -> Dynamic
   -> Eff r ()
-sendMessage px pid message = void (sendMessageChecked px pid message)
+sendMessage px pid message =
+  withFrozenCallStack $ void (sendMessageChecked px pid message)
 
 -- | Send a message to a process addressed by the 'ProcessId'.
 -- See 'SendMessage'. Return @True@ if the process existed.
@@ -210,7 +211,7 @@ sendMessageChecked
   -> Dynamic
   -> Eff r Bool
 sendMessageChecked _ pid message =
-  executeAndResume (SendMessage pid $! message)
+  withFrozenCallStack $ executeAndResume (SendMessage pid $! message)
 
 -- | Send a message to a process addressed by the 'ProcessId'.
 -- See 'SendMessage'.
@@ -221,7 +222,7 @@ sendMessageAs
   -> ProcessId
   -> o
   -> Eff r ()
-sendMessageAs px pid = sendMessage px pid . toDyn
+sendMessageAs px pid = withFrozenCallStack $ sendMessage px pid . toDyn
 
 -- | Exit a process addressed by the 'ProcessId'.
 -- See 'SendShutdown'.
@@ -231,7 +232,7 @@ sendShutdown
   => SchedulerProxy q
   -> ProcessId
   -> Eff r ()
-sendShutdown px pid = void (sendShutdownChecked px pid)
+sendShutdown px pid = withFrozenCallStack $ void (sendShutdownChecked px pid)
 
 -- | Like 'sendShutdown', but also return @True@ iff the process to exit exists.
 sendShutdownChecked
@@ -240,7 +241,8 @@ sendShutdownChecked
   => SchedulerProxy q
   -> ProcessId
   -> Eff r Bool
-sendShutdownChecked _ pid = executeAndResume (SendShutdown pid)
+sendShutdownChecked _ pid =
+  withFrozenCallStack $ executeAndResume (SendShutdown pid)
 
 -- | Start a new process, the new process will execute an effect, the function
 -- will return immediately with a 'ProcessId'.
@@ -249,7 +251,7 @@ spawn
    . (HasCallStack, SetMember Process (Process q) r)
   => Eff (Process q ': q) ()
   -> Eff r ProcessId
-spawn child = executeAndResume (Spawn @q child)
+spawn child = withFrozenCallStack $ executeAndResume (Spawn @q child)
 
 -- | Like 'spawn' but return @()@.
 spawn_
@@ -257,7 +259,7 @@ spawn_
    . (HasCallStack, SetMember Process (Process q) r)
   => Eff (Process q ': q) ()
   -> Eff r ()
-spawn_ = void . spawn
+spawn_ = withFrozenCallStack $ void . spawn
 
 -- | Block until a message was received.
 receiveMessage
@@ -265,7 +267,7 @@ receiveMessage
    . (HasCallStack, SetMember Process (Process q) r)
   => SchedulerProxy q
   -> Eff r Dynamic
-receiveMessage _ = executeAndResume ReceiveMessage
+receiveMessage _ = withFrozenCallStack $ executeAndResume ReceiveMessage
 
 -- | Receive and cast the message to some 'Typeable' instance.
 receiveMessageAs
@@ -273,7 +275,7 @@ receiveMessageAs
    . (HasCallStack, Typeable a, SetMember Process (Process q) r)
   => SchedulerProxy q
   -> Eff r a
-receiveMessageAs px = do
+receiveMessageAs px = withFrozenCallStack $ do
   messageDynamic <- receiveMessage px
   let castAndCheck dm = case fromDynamic dm of
         Nothing -> Left ("Invalid message type received: " ++ show dm)
@@ -302,7 +304,7 @@ self
   :: (HasCallStack, SetMember Process (Process q) r)
   => SchedulerProxy q
   -> Eff r ProcessId
-self _px = executeAndResume SelfPid
+self _px = withFrozenCallStack $ executeAndResume SelfPid
 
 -- | Exit the process.
 exitNormally
@@ -310,7 +312,7 @@ exitNormally
    . (HasCallStack, SetMember Process (Process q) r)
   => SchedulerProxy q
   -> Eff r a
-exitNormally _ = send (Shutdown @q)
+exitNormally _ = withFrozenCallStack $ send (Shutdown @q)
 
 -- | Exit the process with an error.
 exitWithError
@@ -319,7 +321,7 @@ exitWithError
   => SchedulerProxy q
   -> String
   -> Eff r a
-exitWithError _ = send . (ExitWithError @q $!)
+exitWithError _ = withFrozenCallStack $ send . (ExitWithError @q $!)
 
 -- | Thrown an error, can be caught by 'catchRaisedError'.
 raiseError
@@ -328,7 +330,7 @@ raiseError
   => SchedulerProxy q
   -> String
   -> Eff r b
-raiseError _ = send . (RaiseError @q $!)
+raiseError _ = withFrozenCallStack $ send . (RaiseError @q $!)
 
 -- | Catch and handle an error raised by 'raiseError'. Works independent of the
 -- handler implementation.
@@ -339,7 +341,7 @@ catchRaisedError
   -> (String -> Eff r w)
   -> Eff r w
   -> Eff r w
-catchRaisedError _ onErr = interpose return go
+catchRaisedError _ onErr = withFrozenCallStack $ interpose return go
  where
   go :: forall b . Process q b -> (b -> Eff r w) -> Eff r w
   go (RaiseError emsg) _k = onErr emsg
@@ -352,7 +354,8 @@ ignoreProcessError
   => SchedulerProxy q
   -> Eff r a
   -> Eff r (Either String a)
-ignoreProcessError px = catchRaisedError px (return . Left) . fmap Right
+ignoreProcessError px =
+  withFrozenCallStack $ catchRaisedError px (return . Left) . fmap Right
 
 -- | Each process is identified by a single process id, that stays constant
 -- throughout the life cycle of a process. Also, message sending relies on these